requirements.txt
Flask
Flask-MySQL
mysql-connector-python
Flask-WTF
Flask-Session
flask-mysqldb